modified files

+++ src/BonusMilesService.java
@@ -1,5 +1,5 @@
 public class BonusMilesService {
-    public long calculate (int price){
+    public int calculate (int price){
         int mileOne = 20;
         int miles = price/mileOne;
         return miles;
